refactor: Cleanup and readability
