+++ src/Uno.UI.Runtime.Skia.Win32.Support/NativeMethods.txt
@@ -115,6 +115,7 @@ MARGINS
 DWM_WINDOW_CORNER_PREFERENCE
 
 // Functions
+AdjustWindowRectEx
 AdjustWindowRectExForDpi
 AddClipboardFormatListener
 BeginPaint
